refactor(SampleIndexModels): #408 lift Sample.Search.Query + Sample.Search.Result value types out of Services
Two pure value types previously declared inside
`Sources/Services/ReadCommands/Sample.Search.Service.swift` move to
the foundation-layer `SampleIndexModels` target. The
`Sample.Search.Service` actor (the behavioural surface) stays in
Services because it holds an `any Sample.Index.Reader`.

Same lift pattern established in #475 (Sample.Index.Reader),
#465 (Search.Database), #479-#483 (CorePackageIndexingModels family),
and the SearchModels closure-typealias family.

## What moved

- `Sample.Search.Query` (struct, 6 fields) → SampleIndexModels
- `Sample.Search.Result` (struct, 2 fields + 2 computed) → SampleIndexModels

`Sample.Search.Result` references `Sample.Index.Project` and
`Sample.Index.FileSearchResult` which already live in
SampleIndexModels — zero new transitive deps.

## Why

Callers that need to construct a `Sample.Search.Query` (the MCP
`SearchToolProvider` is the main one) or hold a `Sample.Search.Result`
(future CLI / MCP read-paths) can now do so by importing
`SampleIndexModels` instead of pulling in the full `Services` target.
Step toward closing #408.

## Verification

xcrun swift build: clean
xcrun swift test: 1441 tests in 161 suites pass

No source-compatibility break: `Sample.Search.Query` and
`Sample.Search.Result` are still reachable by the same
fully-qualified name; only the target they live in changed.
SearchToolProvider already imports SampleIndexModels (from #475),
so no consumer-side updates required for this PR.
