+++ Packages/Sources/Services/ReadCommands/Sample.Search.Service.swift
@@ -4,64 +4,13 @@ import SampleIndexModels
 import SharedConstants
 import SharedCore
 
-// MARK: - Sample Search Query
+// MARK: - Sample Search
 
-/// Query parameters for sample code searches
-extension Sample.Search {
-    public struct Query: Sendable {
-        public let text: String
-        public let framework: String?
-        public let searchFiles: Bool
-        public let limit: Int
-        /// Optional platform filter (#233). When set together with
-        /// `minVersion`, restricts results to projects whose
-        /// `min_<platform>` column is non-NULL and lex-≤ the requested
-        /// version. nil on either disables the filter.
-        public let platform: String?
-        public let minVersion: String?
-
-        public init(
-            text: String,
-            framework: String? = nil,
-            searchFiles: Bool = true,
-            limit: Int = Shared.Constants.Limit.defaultSearchLimit,
-            platform: String? = nil,
-            minVersion: String? = nil
-        ) {
-            self.text = text
-            self.framework = framework
-            self.searchFiles = searchFiles
-            self.limit = min(limit, Shared.Constants.Limit.maxSearchLimit)
-            self.platform = platform
-            self.minVersion = minVersion
-        }
-    }
-}
-
-// MARK: - Sample Search Result
-
-/// Combined result from project and file searches
-extension Sample.Search {
-    public struct Result: Sendable {
-        public let projects: [Sample.Index.Project]
-        public let files: [Sample.Index.FileSearchResult]
-
-        public init(projects: [Sample.Index.Project], files: [Sample.Index.FileSearchResult]) {
-            self.projects = projects
-            self.files = files
-        }
-
-        /// Check if the result is empty
-        public var isEmpty: Bool {
-            projects.isEmpty && files.isEmpty
-        }
-
-        /// Total count of results
-        public var totalCount: Int {
-            projects.count + files.count
-        }
-    }
-}
+// `Sample.Search.Query` and `Sample.Search.Result` lifted to the
+// `SampleIndexModels` target so callers (`SearchToolProvider` and
+// future MCP / CLI surfaces) can construct queries without importing
+// the full `Services` target. The actor below stays here because it
+// holds behaviour over `any Sample.Index.Reader`.
 
 // MARK: - Sample Search Service
